AngelOneRunner/requirements.txt
playwright>=1.58.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-dotenv>=1.0.0
google-cloud-storage>=2.0.0
